OK, so I got it to cross compile from linux to windows using mingw64 The reason I used this is that working with mingw64 on linux is much more faster, and it is 64 bit.
Most likely, the original one done by My good nephew was a 32 bit build.

this has to be 64 bit for faster graphics and processing.
I did a 32 bit of windows 64 just to try - it did not work very well.

This version just runs the original game without raylib.  about to add raylib